requirements.txt
transformers~=4.47.1
datasets
python-dotenv~=1.0.1
langchain_core
langchain_community
langchain_openai
streamlit~=1.39.0
langserve
groq
bs4
pypdf
sse_starlette
chromadb
langchain-groq
langchainhub
unstructured[all-docs]~=0.16.11
pytesseract
nltk
tiktoken
sentence-transformers
langchain-huggingface
langchain-chroma
langchain~=0.3.13
torch~=2.5.1
pdfplumber~=0.11.4